move output types
