+++ packages/store/ts/config/v2/index.ts
@@ -5,3 +5,4 @@ export * from "./tableShorthand";
 export * from "./tableFull";
 export * from "./table";
 export * from "./store";
+export * from "./output";
