+++ packages/store/ts/config/v2/scope.ts
@@ -1,8 +1,6 @@
 import { Dict, evaluate } from "@arktype/util";
 import { SchemaInput } from "./schema";
-import { StaticAbiType, DynamicAbiType, schemaAbiTypes } from "@latticexyz/schema-type";
-
-export type AbiType = StaticAbiType | DynamicAbiType;
+import { SchemaAbiType as AbiType, StaticAbiType, schemaAbiTypes } from "@latticexyz/schema-type";
 
 export const EmptyScope = { types: {} } as const satisfies ScopeOptions;
 export type EmptyScope = typeof EmptyScope;
