+++ packages/store/ts/config/v2/store.test.ts
@@ -1,5 +1,6 @@
 import { describe, it } from "vitest";
-import { ResolvedStoreConfig, resolveStoreConfig } from "./store";
+import { resolveStoreConfig } from "./store";
+import { ResolvedStoreConfig } from "./output";
 import { attest } from "@arktype/attest";
 import { Hex } from "viem";
